task 2

+++ src/main/java/com/musobek/clickapp/repo/WorkspacePermissionRepository.java
@@ -3,5 +3,5 @@ package com.musobek.clickapp.repo;
 import com.musobek.clickapp.entity.WorkspacePermission;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkspacePermissionRepository extends JpaRepository<WorkspacePermission, Integer> {
+public interface WorkspacePermissionRepository extends JpaRepository<WorkspacePermission, Long> {
 }
